Fix: Remove support for PyPy and library coverage
The `numpy` library failed to build on the PyPy build machine.
Therefore, we have removed support for PyPy (it was unknown if it would
work to begin with).

In addition, changed the `coverage` module's settings in order to just
monitor the project's code coverage, and no the coverage of the test
code. This will hopefully eliminate the extraneous coverage monitoring
of third party libraries, which was not being experienced using the
previous settings on a Windows machine.