Add: CMU Pronouncing Dictionary
The CMU dictionary allows us to count the syllables of words that are in
the dictionary.

> The Carnegie Mellon University Pronouncing Dictionary is a
> machine-readable pronunciation dictionary for North American English
> that contains over 125,000 words and their transcriptions.

While this method does not work for words that are not in the
dictionary, that point may be moot if our (future) word generation
algorithm only uses dictionary words (possibly from WordNet). In the
mean time, this closes gh-1.

For more information, see the [CMU dictionary][] website, and the
[NLTK docs][].

[CMU dictionary]: http://www.speech.cs.cmu.edu/cgi-bin/cmudict
[NLTK docs]: http://nltk.org/_modules/nltk/corpus/reader/cmudict.html