Fix: Install NLTK data during package load
Previously we had been relying on the unit tests to install NLTK
dependencies. For some unknown reason, the `setup_package` method failed
to run when expected, which caused the NLTK dependencies not to be
loaded for some of the tests that required them. Fortunately, this
allowed us to make sure that the module was capable of downloading
missing dependencies before running.